First pass at reshuffling all the code.
