+++ package.json
@@ -43,6 +43,7 @@
 		"@wordpress/edit-widgets": "file:packages/edit-widgets",
 		"@wordpress/editor": "file:packages/editor",
 		"@wordpress/element": "file:packages/element",
+		"@wordpress/embed-block": "file:packages/embed-block",
 		"@wordpress/escape-html": "file:packages/escape-html",
 		"@wordpress/format-library": "file:packages/format-library",
 		"@wordpress/hooks": "file:packages/hooks",
