+++ packages/block-library/package.json
@@ -40,6 +40,7 @@
 		"@wordpress/dom": "file:../dom",
 		"@wordpress/editor": "file:../editor",
 		"@wordpress/element": "file:../element",
+		"@wordpress/embed-block": "file:../embed-block",
 		"@wordpress/escape-html": "file:../escape-html",
 		"@wordpress/i18n": "file:../i18n",
 		"@wordpress/icons": "file:../icons",
