+++ packages/block-library/src/editor.scss
@@ -7,7 +7,7 @@
 @import "./code/editor.scss";
 @import "./columns/editor.scss";
 @import "./cover/editor.scss";
-@import "./embed/editor.scss";
+@import "../embed-block/src/editor.scss";
 @import "./file/editor.scss";
 @import "./classic/editor.scss";
 @import "./gallery/editor.scss";
