+++ packages/block-library/src/embed/index.js
@@ -2,14 +2,13 @@
  * Internal dependencies
  */
 import { common as commonEmbeds, others as otherEmbeds } from './core-embeds';
-import { embedContentIcon } from './icons';
-import { getEmbedBlockSettings } from './settings';
 
 /**
  * WordPress dependencies
  */
 import { __, _x } from '@wordpress/i18n';
 import { createBlock } from '@wordpress/blocks';
+import { embedContentIcon, getEmbedBlockSettings } from '@wordpress/embed-block';
 
 export const name = 'core/embed';
