+++ packages/block-library/src/style.scss
@@ -5,7 +5,7 @@
 @import "./categories/style.scss";
 @import "./columns/style.scss";
 @import "./cover/style.scss";
-@import "./embed/style.scss";
+@import "../embed-block/src/style.scss";
 @import "./file/style.scss";
 @import "./gallery/style.scss";
 @import "./image/style.scss";
