+++ packages/block-library/src/theme.scss
@@ -1,6 +1,6 @@
 @import "./audio/theme.scss";
 @import "./code/theme.scss";
-@import "./embed/theme.scss";
+@import "../embed-block/src/theme.scss";
 @import "./gallery/theme.scss";
 @import "./image/theme.scss";
 @import "./pullquote/theme.scss";
