+++ packages/embed-block/CHANGELOG.md
@@ -1,0 +1,3 @@
+## Master
+
+Initial release.
