+++ packages/embed-block/README.md
@@ -1,0 +1,16 @@
+# Embed Block
+
+Embed block module for WordPress.
+
+## Installation
+
+Install the module
+
+```bash
+npm install @wordpress/embed-block --save
+```
+
+_This package assumes that your code will run in an **ES2015+** environment. If you're using an environment that has limited or no support for ES2015+ such as lower versions of IE then using [core-js](https://github.com/zloirock/core-js) or [@babel/polyfill](https://babeljs.io/docs/en/next/babel-polyfill) will add support for these methods. Learn more about it in [Babel docs](https://babeljs.io/docs/en/next/caveats)._
+
+## API
+
