+++ packages/embed-block/src/editor.scss
