+++ packages/embed-block/src/embed-loading.js
