+++ packages/embed-block/src/embed-placeholder.js
