+++ packages/embed-block/src/icons.js
@@ -1,0 +1,34 @@
+/**
+ * WordPress dependencies
+ */
+import {
+	Path,
+	Polygon,
+	SVG,
+} from '@wordpress/components';
+
+export const embedContentIcon = (
+	<SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+		<Path d="M0,0h24v24H0V0z" fill="none" />
+		<Path d="M19,4H5C3.89,4,3,4.9,3,6v12c0,1.1,0.89,2,2,2h14c1.1,0,2-0.9,2-2V6C21,4.9,20.11,4,19,4z M19,18H5V8h14V18z" />
+	</SVG>
+);
+export const embedAudioIcon = (
+	<SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+		<Path fill="none" d="M0 0h24v24H0V0z" />
+		<Path d="M21 3H3c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h18c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H3V5h18v14zM8 15c0-1.66 1.34-3 3-3 .35 0 .69.07 1 .18V6h5v2h-3v7.03c-.02 1.64-1.35 2.97-3 2.97-1.66 0-3-1.34-3-3z" />
+	</SVG>
+);
+export const embedPhotoIcon = (
+	<SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+		<Path d="M0,0h24v24H0V0z" fill="none" />
+		<Path d="M21,4H3C1.9,4,1,4.9,1,6v12c0,1.1,0.9,2,2,2h18c1.1,0,2-0.9,2-2V6C23,4.9,22.1,4,21,4z M21,18H3V6h18V18z" />
+		<Polygon points="14.5 11 11 15.51 8.5 12.5 5 17 19 17" />
+	</SVG>
+);
+export const embedVideoIcon = (
+	<SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+		<Path d="M0,0h24v24H0V0z" fill="none" />
+		<Path d="m10 8v8l5-4-5-4zm9-5h-14c-1.1 0-2 0.9-2 2v14c0 1.1 0.9 2 2 2h14c1.1 0 2-0.9 2-2v-14c0-1.1-0.9-2-2-2zm0 16h-14v-14h14v14z" />
+	</SVG>
+);
