+++ packages/embed-block/src/index.js
@@ -1,0 +1,6 @@
+/**
+ * Internal dependencies
+ */
+export * from './icons';
+export * from './settings';
+export * from './util';
