+++ packages/embed-block/src/settings.js
