+++ packages/embed-block/src/style.scss
