+++ packages/embed-block/src/theme.scss
